requirements.txt
Django
psycopg2-binary
django-rest-framework
django-versatileimagefield
django-ckeditor
django-cors-headers
django-storages
django-cacheops
django-admin-sortable2
djoser
ipdb
ipython
django-debug-toolbar
django-extensions
django_filters